+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01khx02dxjx6k6awqeqxvr8gmj"
+        "conv_01khxca966851p99k0b9xmm4aq"
     )
     r"""Unique conversation identifier with `conv_` prefix."""