+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHX02EDZZJH9FHPRF52MCENG"
+        "01KHXCA9PHKGFAJ7DFDEANYFFC"
     )
     r"""The unique identifier of the data source"""