+++ packages/orq-rc/RELEASES.md
@@ -5849,3 +5849,13 @@ Based on:
 - [python v4.4.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.17 - packages/orq-rc
+
+## 2026-02-20 11:20:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.18 - packages/orq-rc